Run SendRtpAudio on the worker thread.
This avoids a thread hop to the encoder queue when frame transformer is
used and moves usage of rtp module to the worker thread (although there
is still limited usage on the encoder queue).

A caveat is that this moves frame_encryptor_ usage to the worker thread,
but this should be fine (SRTP runs on the worker thread anyway).

This introduces an extra copy/allocation of the payload buffer in the
non-frame-transformer case.

Bug: none
Change-Id: Ie9ef3e76138473556de1435b10456a95668db7a0
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/432600
Commit-Queue: Jakob Ivarsson‎ <jakobi@webrtc.org>
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46453}