Add Scream traffic policing test
This cl adds a new NetworkBehavior that is implemented as token bucket filter.
The behaviour can either be used as a shaper, if a NetworkQueue is provided,
or as a receive side traffic policer that drops packets if they are received at a too high rate.

A ScreamTest is added that test the behaviour if the link is traffic policed.



Bug: webrtc:447037083
Change-Id: If9ffe5df387977d07f7fd136bde71d88b001a8b2
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/431340
Commit-Queue: Per Kjellander <perkj@webrtc.org>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46446}