Log RTX original sequence number in the RtcEventLog v2.
The RTX OSN is needed for video replay: without it, we will not know
what media packet was recovered from the respective RTX packet.

This CL adds support for RTX OSN to the `RtcEventLog` (v2 only). The
next CL wires this up in `Call`.

Bug: b/305890678, b/423646186
Change-Id: If668ea1bbcb0fb67e1214ca762e18465c5c673f7
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/432502
Commit-Queue: Rasmus Brandt <brandtr@webrtc.org>
Reviewed-by: Björn Terelius <terelius@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46472}