Request refresh frame on resolution scaling changes
When `scale_resolution_down_to` changes in `ConfigureEncoder`, the input
resolution from the source is expected to change. To handle this
cleanly, `RequestRefreshFrame` is now called to prompt the source for a
frame matching the new constraints.

Additionally, the immediate `ReconfigureEncoder` call is deferred if the
resolution scaling has changed. The reconfiguration will instead occur
upon the arrival of the next frame, ensuring the encoder is initialized
with the correct resolution.

Fixed: webrtc:467409350
Change-Id: Icbd19531a45d130460f4763a945d764df3e5e0d2
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/432660
Reviewed-by: Sergey Silkin <ssilkin@webrtc.org>
Commit-Queue: Johannes Kron <kron@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46473}