Remove RtpSenderInternal::SetTransceiverAsStopped
The `is_transceiver_stopped_` flag was being checked in one place and
only set to true when the other flag, `stopped`, is always also set to
true. Checking one flag is enough.

This has the following side effect:

When stopped, CheckSetParameters() will, as before, return an
`INVALID_STATE` error. The error text will now always be "Cannot set
parameters on a stopped sender." as opposed to previously being "Cannot
set parameters on sender of a stopped transceiver" when the transceiver
was *also* stopped.

Bug: none
Change-Id: If35e511b1a23676c359c9baf1445de86c29bc06a
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/434361
Reviewed-by: Henrik Boström <hbos@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46496}