Move FrameInstrumentationGeneratorImpl from api/ to video/
This moves the implementation and unit tests out of the
api/video/corruption_detection and into video/corruption detection -
leaving just an interface sub that delegated to the implementation.

This will make it easier to, in the next step, improve on the frame
sampling implementation.

Bug: webrtc:358039777
Change-Id: I4cf72d010bdd046b1cf4ce985228a0cbab689417
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/434302
Reviewed-by: Per Kjellander <perkj@webrtc.org>
Commit-Queue: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46504}