Bundle media channel teardown with other teardown tasks
This addresses a potential race condition detected here:
https://chromium-swarm.appspot.com/task?id=753fa58d2a45f711&o=true&w=true

The issue was that if PC::Close has not been called and cleanup of state
begins in the destructor, then a pending call to ReportTransportStats()
on the network thread could race with initiated channel cleanup in
RtpTransceiver::GetDeleteChannelWorkerTask().

This CL makes sure that cancellation of the pending ReportTransportStats
task, is run together with the channel cleanup tasks for the
transceivers.

Performance should also marginally be improved since we're doing fewer
hops.

Bug: none
Change-Id: I951052e07376ea55cf1ca562bf4ec5b68f400772
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/434586
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46506}