Fix data race in PeerConnection::ReportTransportStats
This CL fixes a data race detected by TSAN in
PeerConnection::ReportTransportStats. The race occurred because
ReportTransportStats (running on the network thread) was accessing
RtpTransceiver::channel() while the channel was being cleared on the
signaling thread.

The fix resolving the transport name without relying on the
RtpTransceiver's internal channel pointer from the network thread. We
now collect the necessary transport information (MIDs) on the signaling
thread in OnTransportControllerConnectionState and pass it to
ReportTransportStats. ReportTransportStats then uses
JsepTransportController to look up the underlying RtpTransport by MID,
which is safe to do on the network thread.

Bug: webrtc:470709349, webrtc:42222849
Change-Id: If3d4a4d28d87f37420bd514e017b5d49dfce1e50
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/436080
Reviewed-by: Harald Alvestrand <hta@webrtc.org>
Commit-Queue: Tomas Gunnarsson <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46534}