Add FrameSelector utility class, used for ACD frame selection.
This CL adds a class that in a follow-up CL will be wired up to the
automatic corruption detection mechanism. The current implementation
picks a new frame about one second after the previously instrumented
frame. This leads to problem where we never instrument frames that
follow quickly upon keyframes.

This new implementation sets a target timespan after an instrumented
frame. A new cutoff time will be randomly placed within that span,
allowing at least some samples to happen both early and late providing
better temporal coverage.

In addition, we allow the span (and thus inderctly the interval) to be
set differently for low and higher overhead frames. Running the ACD
processing on a frame that is already available in main memory is
cheap, but doing it on a texture-backed frame that requires a download
from GPU to main memory is significantly more expensive. Sampling
with longer intervals mitigates some of this overhead.

Bug: webrtc:358039777
Change-Id: I346f1e4d74166aeb3ed8458adfb698b5265344dd
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/435442
Reviewed-by: Sergey Silkin <ssilkin@webrtc.org>
Commit-Queue: Erik Språng <sprang@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#46533}